Drop miners/composer.py file
